Extracted classes from DataBaseHandler
